+++ internal/table/session.go
@@ -3,6 +3,7 @@ package table
 import (
 	"context"
 	"fmt"
+	"io"
 	"net/url"
 	"strconv"
 	"sync"
@@ -12,6 +13,7 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_Query_V1"
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_Table_V1"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Query"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Table"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_TableStats"
 	"google.golang.org/grpc"
@@ -24,7 +26,7 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/meta"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/operation"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/params"
-	query "github.com/ydb-platform/ydb-go-sdk/v3/internal/query/session"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/session"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/table/config"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/table/scanner"
@@ -40,30 +42,224 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
-// session represents a single table API session.
+type (
+	executor interface {
+		Execute(
+			ctx context.Context,
+			a *allocator.Allocator,
+			request *Ydb_Table.ExecuteDataQueryRequest,
+			callOptions ...grpc.CallOption,
+		) (*transaction, result.Result, error)
+	}
+	tableExecutor struct {
+		client          Ydb_Table_V1.TableServiceClient
+		ignoreTruncated bool
+	}
+	queryExecutor struct {
+		client Ydb_Query_V1.QueryServiceClient
+	}
+)
+
+func statsModeToStatsMode(src Ydb_Table.QueryStatsCollection_Mode) (dst Ydb_Query.StatsMode) {
+	switch src {
+	case Ydb_Table.QueryStatsCollection_STATS_COLLECTION_NONE:
+		return Ydb_Query.StatsMode_STATS_MODE_NONE
+	case Ydb_Table.QueryStatsCollection_STATS_COLLECTION_BASIC:
+		return Ydb_Query.StatsMode_STATS_MODE_BASIC
+	case Ydb_Table.QueryStatsCollection_STATS_COLLECTION_FULL:
+		return Ydb_Query.StatsMode_STATS_MODE_FULL
+	case Ydb_Table.QueryStatsCollection_STATS_COLLECTION_PROFILE:
+		return Ydb_Query.StatsMode_STATS_MODE_PROFILE
+	default:
+		return Ydb_Query.StatsMode_STATS_MODE_UNSPECIFIED
+	}
+}
+
+func txControlToTxControl(src *Ydb_Table.TransactionControl) (dst *Ydb_Query.TransactionControl) {
+	dst = &Ydb_Query.TransactionControl{
+		CommitTx: src.GetCommitTx(),
+	}
+
+	switch t := src.GetTxSelector().(type) {
+	case *Ydb_Table.TransactionControl_BeginTx:
+		switch tt := t.BeginTx.GetTxMode().(type) {
+		case *Ydb_Table.TransactionSettings_SerializableReadWrite:
+			dst.TxSelector = &Ydb_Query.TransactionControl_BeginTx{
+				BeginTx: &Ydb_Query.TransactionSettings{
+					TxMode: &Ydb_Query.TransactionSettings_SerializableReadWrite{
+						SerializableReadWrite: &Ydb_Query.SerializableModeSettings{},
+					},
+				},
+			}
+		case *Ydb_Table.TransactionSettings_SnapshotReadOnly:
+			dst.TxSelector = &Ydb_Query.TransactionControl_BeginTx{
+				BeginTx: &Ydb_Query.TransactionSettings{
+					TxMode: &Ydb_Query.TransactionSettings_SnapshotReadOnly{
+						SnapshotReadOnly: &Ydb_Query.SnapshotModeSettings{},
+					},
+				},
+			}
+		case *Ydb_Table.TransactionSettings_StaleReadOnly:
+			dst.TxSelector = &Ydb_Query.TransactionControl_BeginTx{
+				BeginTx: &Ydb_Query.TransactionSettings{
+					TxMode: &Ydb_Query.TransactionSettings_StaleReadOnly{
+						StaleReadOnly: &Ydb_Query.StaleModeSettings{},
+					},
+				},
+			}
+		case *Ydb_Table.TransactionSettings_OnlineReadOnly:
+			dst.TxSelector = &Ydb_Query.TransactionControl_BeginTx{
+				BeginTx: &Ydb_Query.TransactionSettings{
+					TxMode: &Ydb_Query.TransactionSettings_OnlineReadOnly{
+						OnlineReadOnly: &Ydb_Query.OnlineModeSettings{
+							AllowInconsistentReads: tt.OnlineReadOnly.GetAllowInconsistentReads(),
+						},
+					},
+				},
+			}
+		default:
+			panic(fmt.Sprintf("unknown begin tx settings type: %v", tt))
+		}
+	case *Ydb_Table.TransactionControl_TxId:
+		dst.TxSelector = &Ydb_Query.TransactionControl_TxId{
+			TxId: t.TxId,
+		}
+	default:
+		panic(fmt.Sprintf("unknown tx selector type: %v", t))
+	}
+
+	return dst
+}
+
+func queryExecuteStreamResultToTableResult(
+	ctx context.Context,
+	stream Ydb_Query_V1.QueryService_ExecuteQueryClient,
+) (_ *transaction, _ result.Result, finalErr error) {
+	var (
+		t          *transaction
+		resultSets []*Ydb.ResultSet
+		queryStats *Ydb_TableStats.QueryStats
+	)
+
+	for {
+		if err := ctx.Err(); err != nil {
+			return nil, nil, xerrors.WithStackTrace(err)
+		}
+
+		recv, err := stream.Recv()
+		if err != nil {
+			if xerrors.Is(err, io.EOF) {
+				break
+			}
+
+			return nil, nil, xerrors.WithStackTrace(err)
+		}
+
+		if recv.GetTxMeta() != nil {
+			t = &transaction{
+				Identifier: tx.ID(recv.GetTxMeta().GetId()),
+				control:    table.TxControl(table.WithTxID(recv.GetTxMeta().GetId())),
+			}
+		}
+
+		if recv.GetExecStats() != nil {
+			queryStats = recv.GetExecStats()
+		}
+
+		if rs := recv.GetResultSet(); rs != nil {
+			if idx := int(recv.GetResultSetIndex()); idx == len(resultSets) {
+				resultSets = append(resultSets, recv.GetResultSet())
+			} else if idx < len(resultSets) {
+				resultSets[idx].Rows = append(resultSets[idx].GetRows(), recv.GetResultSet().GetRows()...)
+			} else {
+				return nil, nil, xerrors.WithStackTrace(fmt.Errorf("unexpected result set index: %d", idx))
+			}
+		}
+	}
+
+	return t, scanner.NewUnary(
+		resultSets,
+		queryStats,
+		scanner.WithIgnoreTruncated(false),
+	), nil
+}
+
+func (e queryExecutor) Execute(
+	ctx context.Context,
+	a *allocator.Allocator,
+	executeDataQueryRequest *Ydb_Table.ExecuteDataQueryRequest,
+	callOptions ...grpc.CallOption,
+) (_ *transaction, _ result.Result, finalErr error) {
+	request := a.QueryExecuteQueryRequest()
+
+	request.SessionId = executeDataQueryRequest.GetSessionId()
+	request.ExecMode = Ydb_Query.ExecMode_EXEC_MODE_EXECUTE
+	request.TxControl = txControlToTxControl(executeDataQueryRequest.GetTxControl())
+	request.Query = &Ydb_Query.ExecuteQueryRequest_QueryContent{
+		QueryContent: &Ydb_Query.QueryContent{
+			Syntax: Ydb_Query.Syntax_SYNTAX_YQL_V1,
+			Text:   executeDataQueryRequest.GetQuery().GetYqlText(),
+		},
+	}
+	request.Parameters = executeDataQueryRequest.GetParameters()
+	request.StatsMode = statsModeToStatsMode(executeDataQueryRequest.GetCollectStats())
+	request.ConcurrentResultSets = false
+
+	ctx, cancel := xcontext.WithCancel(xcontext.ValueOnly(ctx))
+	defer cancel()
+
+	stream, err := e.client.ExecuteQuery(ctx, request, callOptions...)
+	if err != nil {
+		return nil, nil, xerrors.WithStackTrace(err)
+	}
+
+	return queryExecuteStreamResultToTableResult(ctx, stream)
+}
+
+func (e tableExecutor) Execute(
+	ctx context.Context,
+	a *allocator.Allocator,
+	request *Ydb_Table.ExecuteDataQueryRequest,
+	callOptions ...grpc.CallOption,
+) (*transaction, result.Result, error) {
+	r, err := executeDataQuery(ctx, e.client, a, request, callOptions...)
+	if err != nil {
+		return nil, nil, xerrors.WithStackTrace(err)
+	}
+
+	return executeQueryResult(r, request.GetTxControl(), e.ignoreTruncated)
+}
+
+var (
+	_ executor = (*tableExecutor)(nil)
+	_ executor = (*queryExecutor)(nil)
+)
+
+// Session represents a single table API session.
 //
 // session methods are not goroutine safe. Simultaneous execution of requests
 // are forbidden within a single session.
 //
 // Note that after session is no longer needed it should be destroyed by
 // Close() call.
-type session struct {
-	onClose   []func(s *session)
+type Session struct {
+	onClose   []func(s *Session)
 	id        string
 	client    Ydb_Table_V1.TableServiceClient
 	status    table.SessionStatus
 	config    *config.Config
+	executor  executor
 	lastUsage atomic.Int64
 	statusMtx sync.RWMutex
 	closeOnce sync.Once
 	nodeID    atomic.Uint32
 }
 
-func (s *session) IsAlive() bool {
+func (s *Session) IsAlive() bool {
 	return s.Status() == table.SessionReady
 }
 
-func (s *session) LastUsage() time.Time {
+func (s *Session) LastUsage() time.Time {
 	return time.Unix(s.lastUsage.Load(), 0)
 }
 
@@ -80,7 +276,7 @@ func nodeID(sessionID string) (uint32, error) {
 	return uint32(id), err
 }
 
-func (s *session) NodeID() uint32 {
+func (s *Session) NodeID() uint32 {
 	if s == nil {
 		return 0
 	}
@@ -96,7 +292,7 @@ func (s *session) NodeID() uint32 {
 	return id
 }
 
-func (s *session) Status() table.SessionStatus {
+func (s *Session) Status() table.SessionStatus {
 	if s == nil {
 		return table.SessionStatusUnknown
 	}
@@ -106,14 +302,14 @@ func (s *session) Status() table.SessionStatus {
 	return s.status
 }
 
-func (s *session) SetStatus(status table.SessionStatus) {
+func (s *Session) SetStatus(status table.SessionStatus) {
 	s.statusMtx.Lock()
 	defer s.statusMtx.Unlock()
 	s.status = status
 }
 
 func newSession(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config) (
-	s *session, finalErr error,
+	s *Session, finalErr error,
 ) {
 	onDone := trace.TableOnSessionNew(config.Trace(), &ctx,
 		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.newSession"),
@@ -129,7 +325,7 @@ func newSession(ctx context.Context, cc grpc.ClientConnInterface, config *config
 	return newTableSession(ctx, cc, config)
 }
 
-func newTableSession(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config) (*session, error) {
+func newTableSession(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config) (*Session, error) {
 	response, err := Ydb_Table_V1.NewTableServiceClient(cc).CreateSession(ctx,
 		&Ydb_Table.CreateSessionRequest{
 			OperationParams: operation.Params(
@@ -149,12 +345,12 @@ func newTableSession(ctx context.Context, cc grpc.ClientConnInterface, config *c
 		return nil, xerrors.WithStackTrace(err)
 	}
 
-	s := &session{
+	s := &Session{
 		id:     result.GetSessionId(),
 		config: config,
 		status: table.SessionReady,
-		onClose: []func(s *session){
-			func(s *session) {
+		onClose: []func(s *Session){
+			func(s *Session) {
 				_, err = s.client.DeleteSession(ctx,
 					&Ydb_Table.DeleteSessionRequest{
 						SessionId: s.id,
@@ -180,29 +376,33 @@ func newTableSession(ctx context.Context, cc grpc.ClientConnInterface, config *c
 			},
 		),
 	)
+	s.executor = tableExecutor{
+		client:          s.client,
+		ignoreTruncated: false,
+	}
 
 	return s, nil
 }
 
-func newQuerySession(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config) (*session, error) {
-	s := &session{
+func newQuerySession(ctx context.Context, cc grpc.ClientConnInterface, config *config.Config) (*Session, error) {
+	s := &Session{
 		config: config,
 		status: table.SessionReady,
 	}
 
-	core, err := query.Open(ctx,
+	core, err := session.Open(ctx,
 		Ydb_Query_V1.NewQueryServiceClient(cc),
-		query.WithConn(cc),
-		query.OnChangeStatus(func(status query.Status) {
+		session.WithConn(cc),
+		session.OnChangeStatus(func(status session.Status) {
 			switch status {
-			case query.StatusClosed:
+			case session.StatusClosed:
 				s.SetStatus(table.SessionClosed)
 				_ = s.Close(context.Background())
-			case query.StatusClosing:
+			case session.StatusClosing:
 				s.SetStatus(table.SessionClosing)
-			case query.StatusInUse:
+			case session.StatusInUse:
 				s.SetStatus(table.SessionBusy)
-			case query.StatusIdle:
+			case session.StatusIdle:
 				s.SetStatus(table.SessionReady)
 			default:
 				s.SetStatus(table.SessionStatusUnknown)
@@ -225,16 +425,19 @@ func newQuerySession(ctx context.Context, cc grpc.ClientConnInterface, config *c
 			},
 		),
 	)
-	s.onClose = []func(s *session){
-		func(s *session) {
+	s.onClose = []func(s *Session){
+		func(s *Session) {
 			_ = core.Close(ctx)
 		},
+	}
+	s.executor = queryExecutor{
+		client: core.Client,
 	}
 
 	return s, nil
 }
 
-func (s *session) ID() string {
+func (s *Session) ID() string {
 	if s == nil {
 		return ""
 	}
@@ -242,9 +445,9 @@ func (s *session) ID() string {
 	return s.id
 }
 
-func (s *session) Close(ctx context.Context) (err error) {
+func (s *Session) Close(ctx context.Context) (err error) {
 	onDone := trace.TableOnSessionDelete(s.config.Trace(), &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).Close"),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).Close"),
 		s,
 	)
 	defer func() {
@@ -270,7 +473,7 @@ func (s *session) Close(ctx context.Context) (err error) {
 	return nil
 }
 
-func (s *session) checkCloseHint(md metadata.MD) {
+func (s *Session) checkCloseHint(md metadata.MD) {
 	for header, values := range md {
 		if header != meta.HeaderServerHints {
 			continue
@@ -284,12 +487,12 @@ func (s *session) checkCloseHint(md metadata.MD) {
 }
 
 // KeepAlive keeps idle session alive.
-func (s *session) KeepAlive(ctx context.Context) (err error) {
+func (s *Session) KeepAlive(ctx context.Context) (err error) {
 	var (
 		result Ydb_Table.KeepAliveResult
 		onDone = trace.TableOnSessionKeepAlive(
 			s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).KeepAlive"),
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).KeepAlive"),
 			s,
 		)
 	)
@@ -328,7 +531,7 @@ func (s *session) KeepAlive(ctx context.Context) (err error) {
 }
 
 // CreateTable creates table at given path with given options.
-func (s *session) CreateTable(
+func (s *Session) CreateTable(
 	ctx context.Context,
 	path string,
 	opts ...options.CreateTableOption,
@@ -407,7 +610,7 @@ func DescribeTable(
 }
 
 // DescribeTable describes table at given path.
-func (s *session) DescribeTable(
+func (s *Session) DescribeTable(
 	ctx context.Context,
 	path string,
 	opts ...options.DescribeTableOption,
@@ -557,7 +760,7 @@ func processChangefeeds(changefeeds []*Ydb_Table.ChangefeedDescription) []option
 }
 
 // DropTable drops table at given path with given options.
-func (s *session) DropTable(
+func (s *Session) DropTable(
 	ctx context.Context,
 	path string,
 	opts ...options.DropTableOption,
@@ -582,7 +785,7 @@ func (s *session) DropTable(
 	return xerrors.WithStackTrace(err)
 }
 
-func (s *session) checkError(err error) {
+func (s *Session) checkError(err error) {
 	if err == nil {
 		return
 	}
@@ -593,7 +796,7 @@ func (s *session) checkError(err error) {
 }
 
 // AlterTable modifies schema of table at given path with given options.
-func (s *session) AlterTable(
+func (s *Session) AlterTable(
 	ctx context.Context,
 	path string,
 	opts ...options.AlterTableOption,
@@ -623,7 +826,7 @@ func (s *session) AlterTable(
 }
 
 // CopyTable creates copy of table at given path.
-func (s *session) CopyTable(
+func (s *Session) CopyTable(
 	ctx context.Context,
 	dst, src string,
 	opts ...options.CopyTableOption,
@@ -690,7 +893,7 @@ func copyTables(
 }
 
 // CopyTables creates copy of table at given path.
-func (s *session) CopyTables(
+func (s *Session) CopyTables(
 	ctx context.Context,
 	opts ...options.CopyTablesOption,
 ) (err error) {
@@ -740,7 +943,7 @@ func renameTables(
 }
 
 // RenameTables renames tables.
-func (s *session) RenameTables(
+func (s *Session) RenameTables(
 	ctx context.Context,
 	opts ...options.RenameTablesOption,
 ) (err error) {
@@ -753,13 +956,13 @@ func (s *session) RenameTables(
 }
 
 // Explain explains data query represented by text.
-func (s *session) Explain(ctx context.Context, sql string) (exp table.DataQueryExplanation, err error) {
+func (s *Session) Explain(ctx context.Context, sql string) (exp table.DataQueryExplanation, err error) {
 	var (
 		result   Ydb_Table.ExplainQueryResult
 		response *Ydb_Table.ExplainDataQueryResponse
 		onDone   = trace.TableOnSessionQueryExplain(
 			s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).Explain"),
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).Explain"),
 			s, sql,
 		)
 	)
@@ -801,14 +1004,14 @@ func (s *session) Explain(ctx context.Context, sql string) (exp table.DataQueryE
 }
 
 // Prepare prepares data query within session s.
-func (s *session) Prepare(ctx context.Context, queryText string) (_ table.Statement, err error) {
+func (s *Session) Prepare(ctx context.Context, queryText string) (_ table.Statement, err error) {
 	var (
 		stmt     *statement
 		response *Ydb_Table.PrepareDataQueryResponse
 		result   Ydb_Table.PrepareQueryResult
 		onDone   = trace.TableOnSessionQueryPrepare(
 			s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).Prepare"),
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).Prepare"),
 			s, queryText,
 		)
 	)
@@ -851,7 +1054,7 @@ func (s *session) Prepare(ctx context.Context, queryText string) (_ table.Statem
 }
 
 // Execute executes given data query represented by text.
-func (s *session) Execute(ctx context.Context, txControl *table.TransactionControl, sql string, params *params.Params,
+func (s *Session) Execute(ctx context.Context, txControl *table.TransactionControl, sql string, params *params.Params,
 	opts ...options.ExecuteDataQueryOption,
 ) (
 	txr table.Transaction, r result.Result, err error,
@@ -892,7 +1095,7 @@ func (s *session) Execute(ctx context.Context, txControl *table.TransactionContr
 
 	onDone := trace.TableOnSessionQueryExecute(
 		s.config.Trace(), &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).Execute"),
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).Execute"),
 		s, q, params,
 		request.QueryCachePolicy.GetKeepInCache(),
 	)
@@ -900,26 +1103,29 @@ func (s *session) Execute(ctx context.Context, txControl *table.TransactionContr
 		onDone(txr, false, r, err)
 	}()
 
-	result, err := s.executeDataQuery(ctx, a, request.ExecuteDataQueryRequest, callOptions...)
+	t, r, err := s.executor.Execute(ctx, a, request.ExecuteDataQueryRequest, callOptions...)
 	if err != nil {
 		return nil, nil, xerrors.WithStackTrace(err)
 	}
 
-	return s.executeQueryResult(result, request.TxControl, request.IgnoreTruncated)
+	if t != nil {
+		t.s = s
+	}
+
+	return t, r, nil
 }
 
 // executeQueryResult returns Transaction and result built from received
 // result.
-func (s *session) executeQueryResult(
+func executeQueryResult(
 	res *Ydb_Table.ExecuteQueryResult,
 	txControl *Ydb_Table.TransactionControl,
 	ignoreTruncated bool,
 ) (
-	table.Transaction, result.Result, error,
+	*transaction, result.Result, error,
 ) {
 	tx := &transaction{
 		Identifier: tx.ID(res.GetTxMeta().GetId()),
-		s:          s,
 	}
 	if txControl.GetCommitTx() {
 		tx.state.Store(txStateCommitted)
@@ -936,8 +1142,9 @@ func (s *session) executeQueryResult(
 }
 
 // executeDataQuery executes data query.
-func (s *session) executeDataQuery(
-	ctx context.Context, a *allocator.Allocator, request *Ydb_Table.ExecuteDataQueryRequest,
+func executeDataQuery(
+	ctx context.Context, client Ydb_Table_V1.TableServiceClient,
+	a *allocator.Allocator, request *Ydb_Table.ExecuteDataQueryRequest,
 	callOptions ...grpc.CallOption,
 ) (
 	_ *Ydb_Table.ExecuteQueryResult,
@@ -948,7 +1155,7 @@ func (s *session) executeDataQuery(
 		response *Ydb_Table.ExecuteDataQueryResponse
 	)
 
-	response, err = s.client.ExecuteDataQuery(ctx, request, callOptions...)
+	response, err = client.ExecuteDataQuery(ctx, request, callOptions...)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
@@ -962,7 +1169,7 @@ func (s *session) executeDataQuery(
 }
 
 // ExecuteSchemeQuery executes scheme query.
-func (s *session) ExecuteSchemeQuery(ctx context.Context, sql string,
+func (s *Session) ExecuteSchemeQuery(ctx context.Context, sql string,
 	opts ...options.ExecuteSchemeQueryOption,
 ) (err error) {
 	request := Ydb_Table.ExecuteSchemeQueryRequest{
@@ -988,7 +1195,7 @@ func (s *session) ExecuteSchemeQuery(ctx context.Context, sql string,
 // DescribeTableOptions describes supported table options.
 //
 //nolint:funlen
-func (s *session) DescribeTableOptions(ctx context.Context) (
+func (s *Session) DescribeTableOptions(ctx context.Context) (
 	desc options.TableOptionsDescription,
 	err error,
 ) {
@@ -1127,14 +1334,14 @@ func (s *session) DescribeTableOptions(ctx context.Context) (
 // via Close() call or fully drained by sequential NextResultSet() calls.
 //
 //nolint:funlen
-func (s *session) StreamReadTable(
+func (s *Session) StreamReadTable(
 	ctx context.Context,
 	path string,
 	opts ...options.ReadTableOption,
 ) (_ result.StreamResult, err error) {
 	var (
 		onDone = trace.TableOnSessionQueryStreamRead(s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).StreamReadTable"),
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).StreamReadTable"),
 			s,
 		)
 		request = Ydb_Table.ReadTableRequest{
@@ -1194,7 +1401,7 @@ func (s *session) StreamReadTable(
 	)
 }
 
-func (s *session) ReadRows(
+func (s *Session) ReadRows(
 	ctx context.Context,
 	path string,
 	keys value.Value,
@@ -1244,7 +1451,7 @@ func (s *session) ReadRows(
 // via Close() call or fully drained by sequential NextResultSet() calls.
 //
 //nolint:funlen
-func (s *session) StreamExecuteScanQuery(ctx context.Context, sql string, parameters *params.Params,
+func (s *Session) StreamExecuteScanQuery(ctx context.Context, sql string, parameters *params.Params,
 	opts ...options.ExecuteScanQueryOption,
 ) (_ result.StreamResult, err error) {
 	var (
@@ -1252,7 +1459,7 @@ func (s *session) StreamExecuteScanQuery(ctx context.Context, sql string, parame
 		q      = queryFromText(sql)
 		onDone = trace.TableOnSessionQueryStreamExecute(
 			s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).StreamExecuteScanQuery"),
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).StreamExecuteScanQuery"),
 			s, q, parameters,
 		)
 		request = Ydb_Table.ExecuteScanQueryRequest{
@@ -1321,14 +1528,14 @@ func (s *session) StreamExecuteScanQuery(ctx context.Context, sql string, parame
 }
 
 // BulkUpsert uploads given list of ydb struct values to the table.
-func (s *session) BulkUpsert(ctx context.Context, table string, rows value.Value,
+func (s *Session) BulkUpsert(ctx context.Context, table string, rows value.Value,
 	opts ...options.BulkUpsertOption,
 ) (err error) {
 	var (
 		a           = allocator.New()
 		callOptions []grpc.CallOption
 		onDone      = trace.TableOnSessionBulkUpsert(s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).BulkUpsert"), s,
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).BulkUpsert"), s,
 		)
 	)
 	defer func() {
@@ -1363,7 +1570,7 @@ func (s *session) BulkUpsert(ctx context.Context, table string, rows value.Value
 }
 
 // BeginTransaction begins new transaction within given session with given settings.
-func (s *session) BeginTransaction(
+func (s *Session) BeginTransaction(
 	ctx context.Context,
 	txSettings *table.TransactionSettings,
 ) (x table.Transaction, err error) {
@@ -1372,7 +1579,7 @@ func (s *session) BeginTransaction(
 		response *Ydb_Table.BeginTransactionResponse
 		onDone   = trace.TableOnTxBegin(
 			s.config.Trace(), &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*session).BeginTransaction"),
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table.(*Session).BeginTransaction"),
 			s,
 		)
 	)