+++ tests/integration/table_truncated_err_test.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"slices"
 	"strconv"
 	"testing"
 
@@ -32,6 +33,22 @@ func TestIssue798TruncatedError(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
+	// clear table
+	{
+		driver.Table().Do(ctx, func(ctx context.Context, s table.Session) error {
+			_, _, err := s.Execute(ctx,
+				table.DefaultTxControl(),
+				fmt.Sprintf("DELETE FROM `%s`;", tablePath),
+				nil,
+			)
+			if err != nil {
+				return err
+			}
+
+			return err
+		}, table.WithIdempotent())
+	}
+
 	// upsert rows
 	{
 		rows := make([]types.Value, rowsLimit)
@@ -41,9 +58,9 @@ func TestIssue798TruncatedError(t *testing.T) {
 				types.StructFieldValue("val", types.TextValue(strconv.Itoa(i))),
 			)
 		}
-		err := driver.Table().Do(ctx, func(ctx context.Context, s table.Session) error {
-			return s.BulkUpsert(ctx, tablePath, types.ListValue(rows...))
-		}, table.WithIdempotent())
+		err := driver.Table().BulkUpsert(ctx, tablePath,
+			table.BulkUpsertDataRows(types.ListValue(rows...)), table.WithIdempotent(),
+		)
 		scope.Require.NoError(err)
 	}
 
@@ -83,6 +100,7 @@ func TestIssue798TruncatedError(t *testing.T) {
 			if count != rowsLimit {
 				return fmt.Errorf("unexpected rows count: %d", count)
 			}
+
 			return rows.Err()
 		}, retry.WithIdempotent(true))
 		scope.Require.NoError(err)
@@ -90,55 +108,86 @@ func TestIssue798TruncatedError(t *testing.T) {
 
 	// upsert 1 row for get 1001 rows and truncated error
 	{
-		err := driver.Table().Do(ctx, func(ctx context.Context, s table.Session) error {
-			return s.BulkUpsert(ctx, tablePath, types.ListValue(types.StructValue(
+		err := driver.Table().BulkUpsert(ctx, tablePath,
+			table.BulkUpsertDataRows(types.ListValue(types.StructValue(
 				types.StructFieldValue("id", types.Int64Value(rowsLimit)),
 				types.StructFieldValue("val", types.TextValue(strconv.Itoa(rowsLimit))),
-			)))
-		}, table.WithIdempotent())
+			))),
+			table.WithIdempotent(),
+		)
 		scope.Require.NoError(err)
 	}
 
 	// select all rows with truncated result error
 	{
-		err := driver.Table().Do(ctx, func(ctx context.Context, s table.Session) error {
-			_, results, err := s.Execute(ctx,
-				table.DefaultTxControl(),
-				fmt.Sprintf("SELECT * FROM `%s`;", tablePath),
-				nil,
-			)
-			if err != nil {
-				return err
-			}
-			if err = results.NextResultSetErr(ctx); err != nil {
-				return fmt.Errorf("no result sets: %w", err)
-			}
-			if results.CurrentResultSet().RowCount() != rowsLimit {
-				return fmt.Errorf("unexpected rows count: %d", results.CurrentResultSet().RowCount())
-			}
-			return results.Err() // expected truncated error
-		}, table.WithIdempotent())
-		scope.Require.ErrorIs(err, result.ErrTruncated)
+		{
+			var rowsCount int
+			err := driver.Table().Do(ctx, func(ctx context.Context, s table.Session) error {
+				_, results, err := s.Execute(ctx,
+					table.DefaultTxControl(),
+					fmt.Sprintf("SELECT * FROM `%s`;", tablePath),
+					nil,
+				)
+				if err != nil {
+					return err
+				}
+				if err = results.NextResultSetErr(ctx); err != nil {
+					return fmt.Errorf("no result sets: %w", err)
+				}
 
-		err = retry.Do(ctx, db, func(ctx context.Context, cc *sql.Conn) error {
-			rows, err := cc.QueryContext(ctx, fmt.Sprintf("SELECT * FROM `%s`;", tablePath))
-			if err != nil {
-				return err
+				rowsCount = results.CurrentResultSet().RowCount()
+				if !slices.Contains([]int{rowsLimit, rowsLimit + 1}, rowsCount) {
+					return fmt.Errorf("unexpected rows count: %d", rowsCount)
+				}
+
+				return results.Err() // expected truncated error
+			}, table.WithIdempotent())
+			switch rowsCount {
+			case rowsLimit:
+				scope.Require.ErrorIs(err, result.ErrTruncated)
+			case rowsLimit + 1:
+				scope.Require.NoError(err)
+			default:
+				scope.Require.Error(err)
+				scope.Require.FailNow("unexpected rows count: %d", rowsCount)
 			}
-			defer func() {
-				_ = rows.Close()
-			}()
-			count := 0
-			for rows.Next() {
-				count++
+		}
+
+		{
+			var rowsCount int
+			err := retry.Do(ctx, db, func(ctx context.Context, cc *sql.Conn) error {
+				rowsCount = 0
+				rows, err := cc.QueryContext(ctx, fmt.Sprintf("SELECT * FROM `%s`;", tablePath))
+				if err != nil {
+					return err
+				}
+				defer func() {
+					_ = rows.Close()
+				}()
+				for rows.Next() {
+					rowsCount++
+				}
+				if !slices.Contains([]int{rowsLimit, rowsLimit + 1}, rowsCount) {
+					return fmt.Errorf("unexpected rows count: %d", rowsCount)
+				}
+
+				return rows.Err()
+			}, retry.WithIdempotent(true))
+			switch driverEngine(db) {
+			case xsql.LEGACY:
+				switch rowsCount {
+				case rowsLimit:
+					scope.Require.ErrorIs(err, result.ErrTruncated)
+				case rowsLimit + 1:
+					scope.Require.NoError(err)
+				default:
+					scope.Require.Error(err)
+					scope.Require.FailNow("unexpected rows count: %d", rowsCount)
+				}
+				scope.Require.ErrorIs(err, result.ErrTruncated)
+			case xsql.PROPOSE:
+				scope.Require.NoError(err)
 			}
-			return rows.Err()
-		}, retry.WithIdempotent(true))
-		switch driverEngine(db) {
-		case xsql.LEGACY:
-			scope.Require.ErrorIs(err, result.ErrTruncated)
-		case xsql.PROPOSE:
-			scope.Require.NoError(err)
 		}
 	}
 
@@ -157,7 +206,7 @@ func TestIssue798TruncatedError(t *testing.T) {
 			if err = results.NextResultSetErr(ctx); err != nil {
 				return fmt.Errorf("no result sets: %w", err)
 			}
-			if results.CurrentResultSet().RowCount() != rowsLimit {
+			if !slices.Contains([]int{rowsLimit, rowsLimit + 1}, results.CurrentResultSet().RowCount()) {
 				return fmt.Errorf("unexpected rows count: %d", results.CurrentResultSet().RowCount())
 			}
 			return results.Err() // expected nil
@@ -182,7 +231,7 @@ func TestIssue798TruncatedError(t *testing.T) {
 			if err = results.NextResultSetErr(ctx); err != nil {
 				return fmt.Errorf("no result sets: %w", err)
 			}
-			if results.CurrentResultSet().RowCount() != rowsLimit {
+			if !slices.Contains([]int{rowsLimit, rowsLimit + 1}, results.CurrentResultSet().RowCount()) {
 				return fmt.Errorf("unexpected rows count: %d", results.CurrentResultSet().RowCount())
 			}
 			return results.Err() // expected nil